Namespace documents routes under case management
