+++ config/locales/es.yml
@@ -58,15 +58,16 @@ es:
         ignore: Si no desea aceptar la invitación, ignore este correo electrónico.
         someone_invited_you: "%{their_name} (%{their_email}) ha invitado a %{your_name} a crear una cuenta en GetYourRefund para ayudar a las personas con sus impuestos."
         subject: Usted ha sido invitado a GetYourRefund
-  documents:
-    edit:
-      title: Editar Documento
-      display_name: Nombre Para Mostrar
-    index:
-      document_type: Clase de documento
-      file_name: Nombre de archivo
-      upload_date: Fecha de carga
-      uploaded_ago: "Hace %{time}"
+  case_management:
+    documents:
+      edit:
+        title: Editar Documento
+        display_name: Nombre Para Mostrar
+      index:
+        document_type: Clase de documento
+        file_name: Nombre de archivo
+        upload_date: Fecha de carga
+        uploaded_ago: "Hace %{time}"
   email:
     user_message:
       subject: Update from GetYourRefund
